practice-12/requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
celery-stubs==0.1.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
Django==4.2
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
frozendict==2.3.7
html5lib==1.1
http-parser==0.9.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
mongo-types==0.15.1
mongoengine==0.27.0
multitasking==0.0.11
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.24.3
pandas==2.0.0
passlib==1.7.4
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.10
sqlparse==0.4.4
tomli==2.0.1
types-psycopg2==2.9.21.9
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
